Rearranged to a modular layout and refactored code to this layout.
