+++ app/config.py
@@ -1,0 +1,18 @@
+# Import the required libraries
+import os
+import html
+from datetime import datetime, time
+import pytz
+from flask import Flask, request, render_template, url_for, redirect, flash
+from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy.sql import func
+from flask_login import LoginManager, UserMixin, login_user, logout_user, login_required, current_user
+from werkzeug.security import generate_password_hash, check_password_hash
+from apscheduler.schedulers.background import BackgroundScheduler
+from apscheduler.triggers.cron import CronTrigger
+import requests
+
+class Config:
+    SQLALCHEMY_DATABASE_URI = "sqlite:///db.sqlite"
+    SQLALCHEMY_TRACK_MODIFICATIONS = False
+    SECRET_KEY = "aXZ!#fl2IlR4ka6n**f6D7e8*F92#D4nf"
